Packaging util functions and fixing typos in the workflow files
