+++ .github/workflows/deploy-apis-to-production.yml
@@ -249,7 +249,7 @@ jobs:
       - name: Deploy to K8S
         uses: Azure/k8s-deploy@v1.4
         with:
-          namespace: "production"
+          namespace: "airflow"
           manifests: |
             k8s/airflow/airflow-app.yaml
           images: "${{ secrets.REGISTRY_URL }}/${{ secrets.PROJECT_ID }}/airqo-apache-airflow:${{ needs.image-tag.outputs.build_id }}"
