+++ .github/workflows/deploy-apis-to-staging.yml
@@ -250,7 +250,7 @@ jobs:
       - name: Deploy to K8S
         uses: Azure/k8s-deploy@v1.4
         with:
-          namespace: "staging"
+          namespace: "stage-airflow"
           manifests: |
             k8s/airflow/stage-airflow-app.yaml
           images: "${{ secrets.REGISTRY_URL }}/${{ secrets.PROJECT_ID }}/airqo-stage-apache-airflow:${{ needs.image-tag.outputs.build_id }}"
