+++ src/airflow/.dockerignore
@@ -2,5 +2,23 @@ venv/
 .idea/
 __pycache__/
 env/
-*.csv
-*.json
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
