+++ src/airflow/.gitignore
@@ -1,5 +1,32 @@
 *.json
 *.csv
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# Environments
+.venv
+env/
+ENV/
+env.bak/
+venv.bak/
 .env
 venv
 venv/
